Major update to texture tinting and parameter access!
Tinting textures is now a core feature, along with scaling and offset. This moves tinting out of Transparency, where it had no business being in the first place. In order to more easily change texture tint during runtime (along with scaling and offset), materials have a BaseParameters shortcut and the Set<T> function takes in an index (that can be used to specify whether your changing tint or whatever for the first, second, third etc. texture)